backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Image Processing
opencv-python==4.9.0.80
Pillow==10.2.0
numpy==1.26.3
scikit-learn==1.4.0
scikit-image==0.22.0

# PDF Generation
reportlab==4.0.9
PyPDF2==3.0.1

# Firebase Admin SDK
firebase-admin==6.4.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Cloud Storage
google-cloud-storage==2.14.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP Client
httpx==0.26.0
aiofiles==23.2.1

# Stripe (Payments)
stripe==7.11.0

# Utilities
python-dotenv==1.0.0
colorama==0.4.6

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2
